+++ projects/pip-webui2-themes/src/lib/themes.module.ts
@@ -1,10 +1,13 @@
-import { NgModule } from '@angular/core';
+import { NgModule, ModuleWithProviders } from '@angular/core';
 
 import { PipThemesService } from './shared/themes.service';
 
-@NgModule({
-  imports: [],
-  declarations: [],
-  providers: [PipThemesService]
-})
-export class PipThemesModule { }
+@NgModule()
+export class PipThemesModule {
+  static forRoot(): ModuleWithProviders {
+    return {
+      ngModule: PipThemesModule,
+      providers: [PipThemesService]
+    };
+  }
+}